Manta: NFC: Fix type-casting bug.
The MSB of the 2-byte length field was truncated
by a wrong type-cast.

This caused incoming NCI packets of over 255 bytes to
be only partially read, causing total confusion in
user-space.

+++ drivers/nfc/bcm2079x-spi.c
@@ -51,7 +51,7 @@
 
 #ifndef STREAM_TO_U16
 #define STREAM_TO_U16(var16, p)	{(var16) = ((u16)(*((u8 *)p+1)) + \
-				((u8)(*((u8 *)p) << 8))); }
+				((u16)(*((u8 *)p) << 8))); }
 #endif
 
 #define STATE_HIGH		1
